VitalFile.to_wfdb: select an appropriate gain and offset.
When exporting signals to WFDB format, we want to preserve the
resolution of the original source values.

If the original sample values are stored as integers (trk.fmt > 2),
then we want to preserve the original samples, and just set the WFDB
gain and baseline accordingly.

If the original sample values are stored as floating-point numbers
(trk.fmt <= 2), then try to guess what the original measurement gain
and baseline might have been.  This uses a simple brute-force search.

Currently this assumes a maximum resolution of 16 bits, as to_wfdb
currently uses 16-bit output format.  It might be desirable to support
higher resolutions in the future, although for higher resolutions the
brute-force method becomes impractical.